requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Async & Job Processing
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# Quantum Computing
qiskit==0.43.3
qiskit-ibm-runtime==0.15.0
pennylane==0.33.1
cirq==1.3.0

# Machine Learning & AI
torch==2.1.1
scikit-learn==1.3.2
numpy==1.26.2
scipy==1.11.4

# Cryptography & Security
cryptography==41.0.7
pyjwt==2.8.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
liboqs-python==0.9.0

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code Quality
black==23.12.0
flake8==6.1.0
isort==5.13.2
bandit==1.7.5
safety==2.3.5

# Utilities
python-dotenv==1.0.0
click==8.1.7
requests==2.31.0
aiohttp==3.9.1

# API Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3
